Added check for ticket master role

+++ listeners/onMessage.js
@@ -17,12 +17,17 @@ module.exports = {
         } else if(message.channelId === process.env.BOOSTSCHANNEL) {
             message.react('♥️');
         } else {
+
             // Automatic "bosshub" reaction
             if(message.content.toLowerCase().includes("bosshub")) {
                 message.react(process.env.BOSSHUBEMOJI);
+
+            // Close ticket command
             } else if(message.content.toLowerCase() == "?closeticket") {
-                message.channel.send("Hello just checking in to make sure your concern been addressed.\n\nIf it has, please close this ticket by clicking the button at the top of this conversation.\n\nIf it hasn't, please respond to this message within 12 hours of your last message or the ticket will be closed and the problem will be considered solved.");
-                message.delete();
+                if(message.member.roles.cache.some(role => role.id === process.env.TICKETMASTERROLE)) {
+                    message.channel.send("Hello just checking in to make sure your concern been addressed.\n\nIf it has, please close this ticket by clicking the button at the top of this conversation.\n\nIf it hasn't, please respond to this message within 12 hours of your last message or the ticket will be closed and the problem will be considered solved.");
+                    message.delete();
+                }
             }
         }
